requirements.txt
langchain
faiss-cpu
langchain_community
openai
azure-data-tables
langchain-openai
streamlit
